backend/Vyapaar-Agent-agents/requirement.txt
google-genai
Pillow
pydantic
pydantic-settings
SQLAlchemy
tabulate
python-dotenv
reportlab